refactor: common 컴포넌트와 payments 컴포넌트 분리
