+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { Outlet } from 'react-router-dom';
 import { CardListProvider } from './context/CardContext';
 import GlobalStyle from './styles/global';
-import Header from './components/Header';
+import PaymentsHeader from './components/payments/PaymentsHeader';
 
 function App() {
   return (
     <CardListProvider>
       <GlobalStyle />
-      <Header />
+      <PaymentsHeader />
       <Outlet />
     </CardListProvider>
   );
