+++ src/stories/Input.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Input from '../components/InputGroup';
+import Input from '../components/common/InputGroup';
 
 const meta: Meta<typeof Input> = {
   title: 'Input',
